refactor: initView View 클래스 내부로 이동
