+++ src/view/View.js
@@ -2,6 +2,8 @@ const RACING_COUNT_HEAD_ID = '#racing-count-head';
 const RACING_COUNT_FORM = '#racing-count-form';
 const RESULT_ID = '#result';
 const CAR_NAMES_INPUT_ID = '#car-names-input';
+const APP_ID = '#app';
+const RESULT_CONTAINER_ID = '#result-container';
 
 export const CAR_NAMES_ERROR = 1;
 export const REPEAT_COUNT_ERROR = 2;
@@ -24,12 +26,21 @@ function hideNextHtml($countHead, $countForm, $result) {
   $resultCopy.style.display = DISPLAY_NONE;
 }
 
+function resetView($appDiv) {
+  const $resultContainer = document.querySelector(RESULT_CONTAINER_ID);
+
+  if ($resultContainer) {
+    $appDiv.removeChild($resultContainer);
+  }
+}
+
 export default class View {
   constructor() {
     this.$countHead = document.querySelector(RACING_COUNT_HEAD_ID);
     this.$countForm = document.querySelector(RACING_COUNT_FORM);
     this.$result = document.querySelector(RESULT_ID);
     this.$carNamesInput = document.querySelector(CAR_NAMES_INPUT_ID);
+    this.$appDiv = document.querySelector(APP_ID);
 
     hideNextHtml(this.$countHead, this.$countForm, this.$result);
   }
@@ -50,5 +61,14 @@ export default class View {
   showNextInput() {
     this.$countHead.style.display = DISPLAY_BLOCK;
     this.$countForm.style.display = DISPLAY_BLOCK;
+  }
+
+  initView() {
+    const $newResultContainer = document.createElement('div');
+    resetView(this.$appDiv);
+
+    $newResultContainer.innerHTML = '';
+    $newResultContainer.id = 'result-container';
+    this.$appDiv.append($newResultContainer);
   }
 }
